refactor: Modularize test_error_handling.rs into scenario-focused modules
Split the 720-line test_error_handling.rs into focused modules:
- catalog_errors.rs: Table, column, schema, and view error tests
- executor_errors.rs: Division by zero, type mismatches, constraints
- storage_errors.rs: Storage layer error tests
- display_and_conversion.rs: Error display and conversion tests

Each module contains related test scenarios, improving code organization
and maintainability. All 27 tests continue to pass (24 passing, 3 ignored).

Closes #699
